remove bookings API route and add users API route with enhanced user data retrieval
